+++ src/model/Subject.java
@@ -1,5 +1,7 @@
 package model;
 
+import type.SubjectType;
+
 public class Subject {
     private String subjectId;
     private String subjectName;
