+++ src/type/Grade.java
@@ -1,4 +1,4 @@
-package model;
+package type;
 
 public enum Grade {
     A, B, C, D, F, N
