+++ src/type/Status.java
@@ -1,4 +1,4 @@
-package model;
+package type;
 
 public enum Status {
     GREEN,
